video_generator/requirements.txt
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3
